+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- concord-console: ability to cancel multiple processes;
 - concord-server: the secret decryption error now contains the
 secret's name;
 - concord-server, concord-console: refresh GitHub webhooks when a